Migrate start and start-ssr commands to mjs
