+++ packages/sku/lib/program/commands/index.mjs
@@ -6,8 +6,8 @@ import lintCommand from './lint/lint.command.js';
 import preCommitCommand from './pre-commit/pre-commit.command.js';
 import serveCommand from './serve/serve.command.js';
 import setupHostsCommand from './setup-hosts/setup-hosts.command.js';
-import startCommand from './start/start.command.js';
-import startSsrCommand from './start-ssr/start-ssr.command.js';
+import { startCommand } from './start/start.command.mjs';
+import { startSsrCommand } from './start-ssr/start-ssr.command.mjs';
 import translationsCommand from './translations/translations.command.js';
 import { buildCommand } from './build/build.command.mjs';
 import { testCommand } from './test/test.command.mjs';
